Set navigation toggle button according to the mockup
On small devices such as Android the nav toggle button is displayed
but cannot be seen because both masthead and bars icon are black.

To fix this, the bars icon color is set to white by changing
the color to "global_Color_light_100" in navToggle.

In addition, show the times icon when sidebar open.