fix(Login): Alert title is not string
Apparently the API returns the error data as an array.
For example, if username is missing the following error is
returned:

{
 ...
 "data": {
  ...
  "username": ["This field may not be blank"]
 }
}

This array is passed to the "title" props in "Alert"
but it is supposed to be a string which ends up with
printing a warning message in the console.

This commit fixes this by converting the array to string
by using the "String" method.

Signed-off-by: Boaz Shuster <boaz.shuster.github@gmail.com>