Added filter results to cost details toolbar
This handles the case where multiple filters are applied.
This handles the case where the page is refreshed or navigated away form.
This handles the case where "Cost By" selection menu changes the page contents.
This introduces patternfly-additions.css, which requires modifications to the existing styles created for the PF3 ListView.