fix(package.json): tslint hack to check test files too
"tsconfig.json" is used by both webpack and tslint.
tslint is using it to determine what files to run lint checking.

If a file is passed to tslint which is excluded tslint prints
an error message saying the file was excluded and quits.

In "tsconfig.json" test files were excluded because they
are not necessary for the UI. However, they are necessary
for linting.

To fix this, I am using two different ts-configs.
"tsconfig.lint.json" for tslint and "tsconfig.json" for webpack.
That way test files can be linted successfully and not be
necessary to run the UI.

Signed-off-by: Boaz Shuster <boaz.shuster.github@gmail.com>