Fixed TypeError in OCP details page
Note that the OCP page still needs some refactoring to use the OCP APIs properly. For example, the toolbar is missing the filter and sort. That will be updated via another PR. This just gets the page working after the work to refactor the OCP dashboard.

Fixes https://github.com/project-koku/koku-ui/issues/337