Add chart legend click functionality to hide/show data
This adds a legend toggle to every chart in the UI. That is, the trend and usage charts on the overview page. The history charts shown in the OCP details. And the pie chart for AWS details.

Also fixed some mismatched chart colors.

Fixes https://github.com/project-koku/koku-ui/issues/391