Harden request errors.
It turns out that a retry limit is not what is needed. The UI makes a request each time the page is updated and it has yet to obtain a report. We simply need to check if we received an error before making another request.

Now if we encounter a 400 error, the unexpected error state is shown. This avoids hitting the server unnecessarily.

Fixes https://github.com/project-koku/koku-ui/issues/590