Historical chart tooltips are cut off
In order to scroll vertically when the window is resized, the chart container cannot overflow. Instead, I've modified our PF4 Victory charts to use voronoiDimension='x', which forces the tooltip to be displayed within the chart's x-axis. This also stacks all tooltips for any given point on the x-axis, which is something we planned to do anyway.

Fixes https://github.com/project-koku/koku-ui/issues/606