fix(onboarding): update ocp and aws providers once on source creation
When adding a new source the App is updated. When that happens
componentDidUpdate is executed where the required behaviour
is to fetch ocp and aws providers if a new source was
added or there are no providers in store yet.

However, in case of "new source was added" it seems like the
condition wasn't good enough ending in sending a lot of requests
to the backend.

The root cause is that when a new source is added the onboarding status
is complete with no errors and the same goes for ocp/aws providers which
makes the condition always true except to the times when the requests are
in progress.

To solve that problem I added another check that when the previous
onboarding status is in progress and now it is complete and there is
no error then update ocp/aws providers.

Signed-off-by: Boaz Shuster <boaz.shuster.github@gmail.com>