fix(costmodels): show available sources in the source step
this fixes the issue where no sources are displayed even though
they exist in the source step.

the issue was due to a change in the API of cost models:
the table was relaying on "provider_uuids" that was replaced
by "providers" and because of that the source list passed to
the table was empty.

Signed-off-by: Boaz Shuster <boaz.shuster.github@gmail.com>