fix(costmodel): make it easy to track unassigned sources in table
Remove the "Default cost model" string from all the sources
that can be assigned. That way the user can easily see that
the ones that are available has no cost model assigned to it
and easier to see that those with a cost model attached to
it cannot be selected.

Signed-off-by: Boaz Shuster <boaz.shuster.github@gmail.com>